Полная переорганизация и часть архитектуры сессий
