+++ api/registration_curl_test.txt
