+++ internal/domain/users.go
@@ -1,0 +1,9 @@
+package domain
+
+type User struct {
+	ID       int    `json:"id"`
+	Email    string `json:"email"`
+	Username string `json:"username"`
+	Password string `json:"password"`
+	Avatar   string `json:"-"`
+}
