+++ internal/models.go
@@ -1,0 +1,10 @@
+package api
+
+/*
+
+type Database struct {
+	echo.Context
+	Users    *[]User
+	Sessions *Sessions
+}
+*/
