+++ internal/users/repository/echo_context/echo_user.go
@@ -1,4 +1,4 @@
-package api
+package echo_context
 
 import (
 	"github.com/labstack/echo"
